pyproject.toml
[tool.poetry]
name = "library-service"
version = "0.1.0"
description = "simple RESTful API for a library management"
authors = ["freddyisman <freddyisman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { extras = ["all"], version = "^0.112.1" }
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.8"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
